+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Cube Cut cube is now oriented based on start face normal
 - More Cube Cut stability (including defensive error messaging when done on invalid faces)
+- Fix version in init!
 
 ## 1.2.0